[Validation] Added input validation
This validation has been added only for the bundle registration command.

It is unnecessary for the package registration, as composer will do it.

This commit brings a new dependency: the Symfony validator component.

It also make the Symfony Form component mandatory in development
environment, in order for the FrameworkBundle to enable the validator
service.